cleaned up some mess
